@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:stuff_pages/enums/searchEnum.dart';
 import 'package:stuff_pages/request/entities/achievement.dart';
 import 'package:stuff_pages/request/entities/game.dart';
 import 'package:stuff_pages/request/http.dart';
@@ -11,7 +12,7 @@ import 'package:stuff_pages/utils/gameUtil.dart';
 import '../navigator.dart';
 
 class ShowTrophy extends StatefulWidget {
-  Game game;
+  late Game game;
 
   ShowTrophy(Game game) {
     this.game = game;
@@ -22,7 +23,7 @@ class ShowTrophy extends StatefulWidget {
 }
 
 class _ShowTrophyState extends State<ShowTrophy> {
-  Game game;
+  late Game game;
   bool donefilter = false;
   List<Achievement> _achievements = [];
   List<Achievement> filteredAchievments = [];
@@ -52,7 +53,7 @@ class _ShowTrophyState extends State<ShowTrophy> {
     filteredAchievments.clear();
     filteredAchievments.addAll(_achievements);
     _achievements.forEach((achievement) {
-      if (!achievement.earned == donefilter) {
+      if (!achievement.earned! == donefilter) {
         filteredAchievments.remove(achievement);
       }
     });
@@ -61,14 +62,15 @@ class _ShowTrophyState extends State<ShowTrophy> {
 
   _getTrophies() {
     filteredAchievments.clear();
-    final endpoint = "achievements/game=" + game.gameId;
+    final endpoint = "achievements/game=" + game.gameId!;
     Api.get(endpoint).then((res) {
       setState(() {
         Iterable list = json.decode(res.body);
         _achievements = list.map((e) => Achievement.fromJson(e)).toList();
-        _achievements.sort((a, b) => a.title.compareTo(b.title));
+        _achievements.sort((a, b) => a.title!.compareTo(b.title!));
         filteredAchievments.addAll(_achievements);
-        donefilter = game.earned != null && (game.earned / game.sum * 100) == 100;
+        donefilter =
+            game.earned != null && (game.earned! / game.sum! * 100) == 100;
         filter();
       });
     });
@@ -88,8 +90,8 @@ class _ShowTrophyState extends State<ShowTrophy> {
     return Scaffold(
       appBar: AppBar(
           backgroundColor: backgroundColor,
-          title: Text(game.title, style: TextStyle(color: fontColor)),
-          actions: <Widget>[doneFilter()]),
+          title: Text(game.title!, style: TextStyle(color: fontColor)),
+          actions: <Widget>[starButton(), doneFilter()]),
       body: Center(
         child: Column(
           children: <Widget>[Expanded(child: _trophyList())],
@@ -114,23 +116,23 @@ class _ShowTrophyState extends State<ShowTrophy> {
             ),
             onDismissed: (direction) {
               Fluttertoast.showToast(
-                  msg: item.title,
+                  msg: item.title!,
                   toastLength: Toast.LENGTH_SHORT,
                   gravity: ToastGravity.BOTTOM,
                   timeInSecForIosWeb: 3,
-                  backgroundColor: item.earned ? deleteColor : addedColor,
+                  backgroundColor: item.earned! ? deleteColor : addedColor,
                   fontSize: 16.0);
               setState(() {
-                item.earned = !item.earned;
+                item.earned = !item.earned!;
                 Api.put('achievements/', item, item.id);
                 filter();
               });
             },
             background:
-                Container(color: item.earned ? deleteColor : addedColor),
+                Container(color: item.earned! ? deleteColor : addedColor),
           ),
           onTap: () {
-            launchURL(game.title + " " + item.title);
+            launchURL(game.title! + " " + item.title!, SearchEnum.youtube);
           },
         );
       },
@@ -138,7 +140,7 @@ class _ShowTrophyState extends State<ShowTrophy> {
   }
 
   Widget getTrophy(Achievement trophy) {
-    final secret = trophy.secret && !trophy.show;
+    final secret = trophy.secret! && !trophy.show!;
     final earned = trophy.earned;
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -152,37 +154,53 @@ class _ShowTrophyState extends State<ShowTrophy> {
                 maxHeight: 200,
               ),
               child: trophyImg(trophy)),
-          title: secret && !earned ? Text(secretTitle, style: TextStyle(color: fontColor)) : Text(trophy.title, style: TextStyle(color: fontColor)),
-          subtitle: secret && !earned
+          title: secret && !earned!
+              ? Text(secretTitle, style: TextStyle(color: fontColor))
+              : Text(trophy.title!, style: TextStyle(color: fontColor)),
+          subtitle: secret && !earned!
               ? Text(secretDescription, style: TextStyle(color: fontColor))
-              : Text(trophy.description, style: TextStyle(color: fontColor)),
+              : Text(trophy.description!, style: TextStyle(color: fontColor)),
           trailing: showButton(trophy),
           onTap: () {
-            launchURL(game.title + " " + trophy.title);
+            launchURL(game.title! + " " + trophy.title!, SearchEnum.youtube);
           },
         ),
       ],
     );
   }
 
-  Widget showButton(Achievement trophy) {
-    if (trophy.secret) {
+  Widget? showButton(Achievement trophy) {
+    if (trophy.secret! && !trophy.earned!) {
       return IconButton(
-          icon: trophy.show
+          icon: trophy.show!
               ? Icon(
-            Icons.lock_open_outlined,
-            color: fontColor,
-          )
+                  Icons.lock_open_outlined,
+                  color: fontColor,
+                )
               : Icon(
-            Icons.lock_outlined,
-            color: fontColor,
-          ),
+                  Icons.lock_outlined,
+                  color: fontColor,
+                ),
           onPressed: () {
             setState(() {
-              trophy.show = !trophy.show;
+              trophy.show = !trophy.show!;
             });
           });
     }
     return null;
+  }
+
+  Widget starButton() {
+    return IconButton(
+        icon: Icon(
+          game.star! ? Icons.star : Icons.star_border,
+          color: futureColor,
+        ),
+        onPressed: () {
+          setState(() {
+            game.star = !game.star!;
+            Api.put("games/", game, game.id);
+          });
+        });
   }
 }
