@@ -1,6 +1,6 @@
 class TrophyNumber {
-  int earned;
-  int sum;
+  int? earned;
+  int? sum;
 
   TrophyNumber({this.earned, this.sum});
 
